+++ NEWS.md
@@ -1,3 +1,13 @@
+# spsUtil 0.2.0 Dev
+
+## New feature
+
+-   `spsOption` has a new argument `.list`: set more than one options a time by passing a list of options and their values to it.
+
+## bug fix
+
+- Fix NA bug in `emptyIsFalse`.
+
 # spsUtil 0.1.2
 
 ## Major change